Add NotificationManager Singleton
